Inventory refactoring

+++ Game/Events/Inventory/IInventoryEvent.cs
@@ -5,6 +5,6 @@ namespace UAlbion.Game.Events.Inventory
     public interface IInventoryEvent
     {
         InventoryType InventoryType { get; }
-        int InventoryId { get; }
+        ushort InventoryId { get; }
     }
 }
